WIP: Weapons carry 2.5
